backend/requirements.txt
fastapi==0.116.1
uvicorn==0.23.2
pydantic==2.6.4
requests==2.32.3
groq==0.31.0  
pdfplumber==0.11.0
python-multipart  # Needed for file uploads in FastAPI
python-pptx==0.6.23
sqlalchemy==2.0.23
python-dotenv==1.0.1